to real

+++ include/type/type.hpp
@@ -24,7 +24,7 @@ struct Light : public cv::RotatedRect {
       : cv::RotatedRect(cv::minAreaRect(contour)) {
 
     center = std::accumulate(contour.begin(), contour.end(), cv::Point2f(0, 0),
-                             [n = static_cast<float>(contour.size())](
+                             [n = static_cast<double>(contour.size())](
                                  const cv::Point2f &a, const cv::Point &b) {
                                return a + cv::Point2f(b.x, b.y) / n;
                              });
@@ -54,7 +54,7 @@ struct Light : public cv::RotatedRect {
   cv::Point2f axis;
   double length;
   double width;
-  float tilt_angle;
+  double tilt_angle;
 };
 struct LightParams {
   // width / height
@@ -127,7 +127,7 @@ typedef struct ArmorObject {
 
   ArmorColor color;
   ArmorNumber number;
-  float prob;
+  double prob;
   std::vector<cv::Point2f> pts;
   std::vector<cv::Point2f> pts_binary;
   cv::Rect box;
@@ -188,8 +188,8 @@ struct Armor {
   tf2::Quaternion ori;
   Position target_pos;
   tf2::Quaternion target_ori;
-  float distance_to_image_center;
-  float yaw;
+  double distance_to_image_center;
+  double yaw;
   std::chrono::steady_clock::time_point timestamp;
 };
 struct Armors {
@@ -207,14 +207,14 @@ struct Target {
 
   Position position_ = Position();
   Position velocity_ = Position();
-  float yaw = 0;
-  float v_yaw = 0;
-  float radius_1 = 0.24;
-  float radius_2 = 0.24;
-  float d_za = 0;
-  float d_zc = 0;
-  float yaw_diff;
-  float position_diff;
+  double yaw = 0;
+  double v_yaw = 0;
+  double radius_1 = 0.24;
+  double radius_2 = 0.24;
+  double d_za = 0;
+  double d_zc = 0;
+  double yaw_diff;
+  double position_diff;
 
   void clear() {
     id = ArmorNumber::UNKNOWN;
@@ -243,11 +243,11 @@ struct SyncedData {
 };
 struct GimbalCmd {
   std::chrono::steady_clock::time_point timestamp;
-  float pitch = 0;
-  float yaw = 0;
-  float yaw_diff = 0;
-  float pitch_diff = 0;
-  float distance = -1;
+  double pitch = 0;
+  double yaw = 0;
+  double yaw_diff = 0;
+  double pitch_diff = 0;
+  double distance = -1;
   bool fire_advice = false;
   int select_id = -1;
 };
